requirements.txt
Django==3.2.7
djangorestframework==3.13.1
drf-yasg==1.21.3
Pillow==9.2.0
django-rest-auth
django-allauth
transliterate
    